java: Add FServlet.Builder.executorService
Submitting to an ExecutorService is an easy way to limit concurrency
differently on a per-endpoint basis rather than relying on the global
shared HTTP thread pool.  Additionally, this also allows the same
concurrent limits to be applied if the service handler is exposed by
multiple servers (for example, NATS and HTTP).  Finally, this also makes
it easy to specify meaningful service thread names rather than using
generic HTTP pooled thread names.  This could be extended in the future
to use the asynchronous servlet mechanism so the HTTP thread can be
reused while the Frugal thread is still processing.