dart: Check primitive fields in validate
Java does not support nullable primitives, so the validate() method has
nothing to do for those fields, so they only need to be validated in the
read() method.  The Dart generator appears to have been copied from the
Java generator, but Dart does support nullable primitives, which means
the write() code path is missing validation.